+++ js/data.js
@@ -1,0 +1,13 @@
+//getComments
+const MIN_INTEGER = 1;
+const MAX_INTEGER = 6;
+
+//getMocks(Likes)
+const MIN_LIKES_INTEGER = 15;
+const MAX_LIKES_INTEGER = 200;
+
+//getMocks(Comments)
+const MIN_COMMENTS_INTEGER = 0;
+const MAX_COMMENTS_INTEGER = 30;
+
+export { MIN_INTEGER, MAX_INTEGER, MIN_LIKES_INTEGER, MAX_LIKES_INTEGER, MIN_COMMENTS_INTEGER, MAX_COMMENTS_INTEGER };
